requirements.txt
newsapi-python
azure.cosmos
pyyaml
openai
transformers
bs4
torch 
torchvision 
torchaudio
requests